% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEISAcompR.R
\name{getEISAcompR}
\alias{getEISAcompR}
\title{getEISAcompR}
\usage{
getEISAcompR(Exons, Introns, design, filterExpr = TRUE, percent = 0.5, cpm = 1)
}
\arguments{
\item{Exons}{Exonic raw counts (genes in rows and samples in columns).}

\item{Introns}{Intronic raw counts (genes in rows and samples in columns).}

\item{design}{Design matrix (1st = sample names; 2nd = group assignment + optionally one additional column with batch effect).}

\item{filterExpr}{Boolean to perform filtering based on expression criteria to remove lowly expressed genes (TRUE/FALSE).}

\item{percent}{Percentage of samples showing minimum expression threshold for filtering (50\% by default).}

\item{cpm}{counts-per-million (CPM) expression threshold for filtering lowly expressed genes (1 CPM by default).}
}
\value{
{object with four tables.
resDE = Differential Expression analysis for exonic counts using glmQLFTest from edgeR package.
resTc = EISA for the transcriptional component effect on each analyzed gene.
resPTc = EISA for the post-transcriptional component effect on each analyzed gene.
Expr_Int = Normalized log2 expression matrix for Intronic counts.
Expr_Ex = Normalized log2 expression matrix for Exonic counts}
}
\description{
This is the main function used to calculate exon/intron split estimates
and compute transcriptional and post-transcriptional components of each gene,
as well as to infer the significance of each regulatory component independently.
The pipeline is implemented for two-group contrast (tipically control vs treated)
and includes the possibility of user-defined batch correction.
Users must prepare their exon/intron counts, design and batch (optional)
matrices including data for same number of selected samples.
In the event that design matrix includes an additional batch
effect apart from sample name (1st) and group assignment (2nd),
it will be considered as independent batch effect.
Further details:
If no batch correction needs to be implemented, please only use a two-column design matrix.
For transcriptional (Tc) and post-transcriptional (PTc) components, generally, the higher their absolute values (either showing negative or positive regulatory influence), the more relevant regulatory effects could be inferred. Please be aware that the abscence of significance in PTc component might indicate the presence of mixed transcriptional and post-transcriptional componentes affecting the same gene. Any PTc component showing significant interaction with any other transcriptional (Tc) influence detected at intronic levels will be shown as not significant. Users should compare canonical differential expression (DE) results and significance for their genes of interest, as well as their Tc and PTc components, in order to extract meaningfull information about the putative regulatory influence affecting their genes of interest.

Output interpretation:
Significant PTc + Significant DE = Gene showing Post-transcriptional regulatory signal
Significant Tc + Significant DE = Gene showing Transcriptional regulatory signal
Non-significant PTc + Significant Tc + Significant DE = Gene showing mixed Transcriptional and Post-transcriptional regulatory signal
}
\examples{
{
\dontrun{
eisa <- getEISAcomp(Exons=exon_counts, Introns=intron_counts, design=design_matrix,
filterExpr=TRUE, percent=0.5, cpm=1)
}

}
}
\references{
\enumerate{
 \item Gaidatzis D et al. (2015) Analysis of intronic and exonic reads in RNA-seq data characterizes transcriptional and post-transcriptional regulation. Nature Biotechnology, 33, 722–729.
 \item Lawrence M et al. (2013) Software for computing and annotating genomic ranges. PLoS Computational Biology, 9, e1003118.
 \item Liao Y et al. (2019) The R package Rsubread is easier, faster, cheaper and better for alignment and quantification of RNA sequencing reads. Nucleid Acids Research, 47, e47.
 \item Robinson MD et al. (2010) edgeR: a Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics, 26, 139–140.
}
}
\author{
Emilio Mármol Sánchez
}
