% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEISAcounts.R
\name{getEISAcounts}
\alias{getEISAcounts}
\title{getEISAcounts}
\usage{
getEISAcounts(files, annotFile, strandness, nthreads, PairedEnd = TRUE)
}
\arguments{
\item{files}{PATH to input BAM/SAM files containing read mapping results.}

\item{annotFile}{PATH to exonic/intronic GTF annotation file.}

\item{strandness}{Strandness. Character. You have to select one out "unstranded","stranded_forward","stranded_reversed".}

\item{nthreads}{Number of available threads for parallel computation.}

\item{PairedEnd}{Boolean specifying if input sequencing data is of type Paired-end (TRUE/FALSE).}
}
\value{
results. It will create a raw count matrix stored at counts object with quantification estimates of each mapped read to the corresponding exonic/intronic regions and gene assignment.
}
\description{
This function allows users to quantify the level of
expression of their aligned reads by making use of custom-created exonic/intronic
GTF annotation files. The featureCounts built-in function from Rsubread package is
used for such purpose. Users may optionally decide to use any other preferred
quantification pipeline provided the use of exonic/intronic GTF annotation
as a reference.
}
\examples{
{
\dontrun{
exon_counts <- getEISAcounts(files=vector_of_input_BAM/SAM, annotFile="PATH_to_exon_GTF",
strandness="unstranded", nthreads=4, PairedEnd=TRUE)
intron_counts <- getEISAcounts(files=vector_of_input_BAM/SAM, annotFile="PATH_to_intron_GTF",
strandness="unstranded", nthreads=4, PairedEnd=TRUE)
}

}
}
\author{
Emilio Mármol Sánchez
}
