% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEISAgtfs.R
\name{makeEISAgtfs}
\alias{makeEISAgtfs}
\title{makeEISAgtfs}
\usage{
makeEISAgtfs(
  annotFile,
  path_temp_files = "~/",
  boundaryFix = 10,
  show_message = TRUE
)
}
\arguments{
\item{annotFile}{PATH to reference GTF annotation file.}

\item{path_temp_files}{Home is the default "~/". This argument indicates the path were the temp_dir with exons, introns info should be created.}

\item{boundaryFix}{Exon junction boundary correction threshold (10 bp by default).}

\item{show_message}{(logical) if TRUE the proccess history is printed. Default TRUE.}
}
\value{
results. It returns a EIsacompR object with the exons and introns information
}
\description{
This is the first function of EISACompR. makeEISAgtfs aims to
build custom annotation files in GTF format for late reads
quantification spanning either exonic or intronic genomic ranges.
Reference GTF annotation for your species of interest is needed in
order to generate the corresponding exonic/intronic regions for
 running EISA analysis.
}
\examples{
{
\dontrun{
GTFS<-makeEISAgtfs(annotFile="path_to_ann_file")
}

}

}
\author{
Emilio Mármol Sánchez
}
