% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEISAgtfs.R
\name{writeEISAgtfs}
\alias{writeEISAgtfs}
\title{writeEISAgtfs}
\usage{
writeEISAgtfs(eisaR.obj, out.path = "~/")
}
\arguments{
\item{eisaR.obj}{EISACompR object previously stored with exon/intron split GTFs.}

\item{out.path}{PATH to desired output. Default is "~/".}
}
\description{
This function allows to export the exon/intron annnotation files in GTF format.
}
\examples{
{
\dontrun{
writeEISAgtfs(GTFs, "~/")
}

}
}
\author{
Emilio Mármol Sánchez
}
